Added Inventory Vendor Requests
A user can now go in the Inventory portion of the application and create
vendor requests.  This functionality was added so that a transfer order
could be initially requested as well as Inventory adjustments.
A How-To document (Created by VanderHey) was also added.  This document
explains what the various pages in the application are for and how a
supplier should interact with them.