Solved issue where all vendor request are initially loaded into memory
I had initially implemented a method which filtered the initial Vendor
Request result set to 100 records. This created a problem when we
reached 101 records in the database. I recoded that portion of the
application so it functions properly and only a default subset of
records are loaded into memory prior to being filtered.